add mcp and change tools to be mcp servers
