+++ tiny_scientist/mcp/__init__.py
@@ -1,0 +1,5 @@
+"""MCP server and implementation package for Tiny Scientist tools."""
+
+from __future__ import annotations
+
+__all__ = []
